requirements.txt
alembic==1.13.3
fastapi==0.115.3
passlib==1.7.4
pydantic==2.9.2
pydantic_settings==2.6.0
python_multipart==0.0.12
python_jose==3.3.0
SQLAlchemy==2.0.36
starlette==0.41.0
uvicorn==0.32.0
bcrypt==4.2.0
psycopg2-binary==2.9.10
requests